SPARK-25003: Add helper methods to create new Extensions from Conf
Previously the only way to add extensions to the session was via the
getOrCreate method of the SparkSession Builder. To facilitate non-scala
Session creation we add a new constructor which takes in just the
context and Extensions. Then we also add a new Extensions constructor
which given a SparkConf generates an Extensions object with user config
already applied.
