+++ python/pyspark/sql/session.py
@@ -212,15 +212,17 @@ class SparkSession(object):
         self._sc = sparkContext
         self._jsc = self._sc._jsc
         self._jvm = self._sc._jvm
+
         if jsparkSession is None:
             if self._jvm.SparkSession.getDefaultSession().isDefined() \
                     and not self._jvm.SparkSession.getDefaultSession().get() \
                         .sparkContext().isStopped():
                 jsparkSession = self._jvm.SparkSession.getDefaultSession().get()
             else:
-                jsparkSession = self._jvm.SparkSession.builder() \
-                    .sparkContext(self._jsc.sc()) \
-                    .getOrCreate()
+                extensions = self._sc._jvm.org.apache.spark.sql\
+                    .SparkSessionExtensions(self._jsc.getConf())
+                jsparkSession = self._jvm.SparkSession(self._jsc.sc(), extensions)
+
         self._jsparkSession = jsparkSession
         self._jwrapped = self._jsparkSession.sqlContext()
         self._wrapped = SQLContext(self._sc, self, self._jwrapped)
